Everything works. Only thing left to do is format output to match ZyLabs. Yasss

+++ Lab2/Lab2/Dealership.h
@@ -14,7 +14,7 @@ public:
 
 	void AddShowroom(const Showroom* s);
 	void ShowInventory() const;
-	unsigned int GetAveragePrice();
+	float GetAveragePrice();
 
 private:
